added missing validation for some components + refactor roles, permissions and policies logic

+++ app/Http/Controllers/PostController.php
@@ -14,6 +14,7 @@ use App\Models\Post;
 use App\Services\ImageConversionService;
 use App\Traits\HandleMediaUploads;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Redirect;
 use Inertia\Inertia;
 use Inertia\Response;
@@ -33,6 +34,8 @@ class PostController extends Controller
 
     public function index(GetIndexData $action): Response
     {
+        Gate::authorize('viewAny', Post::class);
+
         $posts = $action->execute(
             model: Post::class,
             allowedFilters: [
@@ -51,6 +54,8 @@ class PostController extends Controller
 
     public function store(StorePostRequest $request, Create $action): RedirectResponse
     {
+        Gate::authorize('create', Post::class);
+        
         $action->execute(
             PostFormData::from($request->validated()),
             $request->file('thumbnail')
@@ -67,11 +72,15 @@ class PostController extends Controller
 
     public function create(): Response
     {
+        Gate::authorize('create', Post::class);
+
         return Inertia::render('admin/posts/create/Page');
     }
 
     public function update(UpdatePostRequest $request, Update $action, Post $post): RedirectResponse
     {
+        Gate::authorize('update', $post);
+
         $action->execute(
             $post,
             PostFormData::from($request->validated()),
@@ -92,6 +101,8 @@ class PostController extends Controller
 
     public function edit(Post $post): Response
     {
+        Gate::authorize('update', $post);
+
         $post->load('tags');
 
         return Inertia::render('admin/posts/edit/Page', [
@@ -101,6 +112,8 @@ class PostController extends Controller
 
     public function destroy(Post $post): RedirectResponse
     {
+        Gate::authorize('forceDelete', $post);
+
         $post->forceDelete();
 
         $currentPage = request('currentPage');
@@ -123,6 +136,8 @@ class PostController extends Controller
 
     public function soft_delete(Post $post): RedirectResponse
     {
+        Gate::authorize('delete', $post);
+
         $currentPage = request('currentPage');
         $post->published_at = null;
         $post->save();
@@ -143,6 +158,8 @@ class PostController extends Controller
 
     public function restore(Post $post): RedirectResponse
     {
+        Gate::authorize('restore', $post);
+
         $post->restore();
 
         Inertia::flash('toast',
